Fix ambiguous function call of ceil

+++ sim/include/texture.hpp
@@ -1,5 +1,8 @@
 // Original code (c) 2023 Alan Jian
 // Licensed under MIT License
+//
+// Modifications (c) 2025 jin11109
+// Licensed under MIT License
 
 #pragma once
 
@@ -15,7 +18,7 @@
 extern const uint32_t TEXTURE[TEXTURE_WIDTH * TEXTURE_HEIGHT];
 
 inline RGB8 sample_texture(Vec2f uv) {
-    int x = ceil(uv.x * (TEXTURE_WIDTH - 1));
-    int y = ceil(uv.y * (TEXTURE_HEIGHT - 1));
+    int x = ceil((double)uv.x * (TEXTURE_WIDTH - 1));
+    int y = ceil((double)uv.y * (TEXTURE_HEIGHT - 1));
     return RGB8::decode(TEXTURE[y * TEXTURE_WIDTH + x]);
 }
